audio dataset refactor
