+++ clip_benchmark/datasets/audio/gtzan.py
@@ -86,11 +86,7 @@ class GTZAN(Dataset):
             print(f"Warning: Failed to load {file_path}, returning silence.")
         
         # Pad/Crop
-        if len(audio_data) < self.target_len:
-            padding = self.target_len - len(audio_data)
-            audio_data = np.pad(audio_data, (0, padding), 'constant')
-        else:
-            audio_data = audio_data[:self.target_len]
+        audio_data = utils.pad_or_crop(audio_data, self.target_len)
         
         # Convert to PyTorch Tensor
         audio_tensor = torch.from_numpy(audio_data).float()
